+++ commands/settings/help.js
@@ -3,7 +3,7 @@ module.exports = {
     description: "Understand how the bot's settings works",
     needPerms: false,
     run: async (e, vars, args, settings) => {
-        let embed = new vars.discord.MessageEmbed()
+        let embed = vars.newEmbed()
         .setTitle('How it works ?')
         .setDescription(`
         > First, generate a channel with \`${settings.prefix}settings generate\` (you can also use \`${settings.prefix}settings set <channel>\`)
@@ -15,7 +15,6 @@ module.exports = {
 
         ||Don't forget: the setting's name is before the "=" and the value of this setting is after the "=" and inside quotes!||
         `)
-        .setColor('RANDOM')
         e.reply(embed)
     }
 }